✨ Refactor folder structure
