+++ server/server.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 const dotenv = require("dotenv");
-const app = require("./app");
+const app = require("./src/app");
 
 // DotENV
 dotenv.config({ path: "./.env" });
